Improve horizontal zooming
Zomming is now done with `this.scaleX * (1 + scaleXDelta)`.
This makes the `scaleXDelta` relative and independent of the former zoom level (a value of 0.1 will alway show 10% more of the piano roll)

The mouse position is now retrived with `e.nativeEvent.offsetX` which is more general and should work if more content is added to the left side.

The platform check is also removed since it is not needed.