Feat: Add ghost track button to track list
As default a track will now be a ghost track. Personally, I think I prefer it this way, but it can easily be changed.

This commit also removes the ghost track drop down since it now operates globally (not a separate ghost track for each track). The former flow was more cusomizeable, but too cumbersome to use.